Lots of bug fixes to get Lyrics3Test case created and working

+++ src/main/java/org/farng/mp3/id3/FrameBodyCOMM.java
@@ -103,7 +103,7 @@ public class FrameBodyCOMM extends AbstractID3v2FrameBody {
     }
 
     public byte getTextEncoding() {
-        return ((Byte) getObject(ObjectNumberHashMap.TEXT_ENCODING)).byteValue();
+        return (byte) (long) (Long) getObject(ObjectNumberHashMap.TEXT_ENCODING);
     }
 
     protected void setupObjectList() {
